+++ react/typings/sbdsefuat.shopping-cart.d.ts
@@ -1,0 +1,1 @@
+declare module 'sbdsefuat.shopping-cart/OrderFormCustomFieldHooks'